+++ huaweicloud/services/dbss/data_source_huaweicloud_dbss_flavors.go
@@ -170,7 +170,7 @@ func flattenGetFlavorsResponseBodyFlavor(resp interface{}) []interface{} {
 	curArray := curJson.([]interface{})
 	rst := make([]interface{}, 0, len(curArray))
 	for _, v := range curArray {
-		memory := utils.PathSearch("ram", v, 0).(float64) / 1024
+		memory := utils.PathSearch("ram", v, float64(0)).(float64) / 1024
 		rst = append(rst, map[string]interface{}{
 			"id":                 utils.PathSearch("id", v, nil),
 			"level":              utils.PathSearch("level", v, nil),
